Fix Makefile, remove unneeded sudo, fix inspvasa, fix command format.
Makefile in root now calls pybin Makefile. The user was added to the dialout
group, so now /dev/ttyUSB0 can be read without sudo. For this reason, sudo was
removed from sensors.sh, which now writes log files with more reasonable perms.

INSPVASA in span.py was modified to handle binary, the name was changed to
reflect that it is not just an ASCII command. The message writing was sent
through the generic messsage sending functions.

Fixed pybin to properly quote the struct packing format for marktime and
mark2time.