Use image queue.
The Image queue allows image access in FIFO format, so that frames are not
dropped due to I/O bottlenecks. In the previous method, the most recently
written frame was processed, and so some frames were skipped and lost forever.

Additionally, the memory buffer was extended to 64 frames per camera, however it
is not known at this time if this is a good amount of memory.