Modify output format and add offset argument.
The previous output had a path to my computer hardcoded. The current output just
prints the frame number and the user can `sed` the proper path in.

The -o argument allows the user to set the initial offset between the camera and
gps. This usage is preferred because when the initial offset is calculated in a
controlled manner. The default scenario assumes the first line of the gps maps
to the first line of the camera, which is certainly not a good assumption.