Update usage. Fix memory leak. Add one-time mode.
Added usage message to top to describe new arguments. Users can now use -d for a
debug mode where images are not saved and event timeouts are not set. There is
also -1 for one-time mode which closes the program after the first frame.

There was a memory leak where cameras were not freed in the signal handler and
this was fixed.