Modify syncTime input and output formats.
The GPS input is now provided in seconds only. The user is responsible for
handling the GPS week beforehand. The output is now in GPS time. This is done to
match the rest of the SLAM input and relieves this program from doing any UTC
conversions whatsoever.

Camera input usage message explicitly specifies nanoseconds.