Timestamps in GPS seconds only.
Wrote a function to convert gpswk,sec to the total number of seconds. This makes
removes the burden on downstream programs from having to know what gpswk is. In
this way, they can handle either UTC or GPS seconds, or any other seconds based
clock notion in the same manner.