Use camtime for syncTime timestamp.
For the time being, we cannot get timestamps with better than millisecond
resolution from mark2timeb. The cam time give nanosecond resolution, so we are
using this for the timestamp output. If we can get high resolution from the
mark2timeb command, we should revert to this in order to use a single clock for
all sensors.