Fix quaternion calculation and write output to file.
euler2qbw() expects attitude in radians, but the span outputs in degrees. The
conversion from degrees to radians was added into the function.

When running live, there were issues with the output being written to stdout
that don't appear when writing directly to a file, so we are now writing to a
timestamped file.