Workaround for header length issue.
The Novatel SPAN-IGM-A1 does not return header messages of fixed length as
described in the documentation. Nor does the 'header length' byte correspond to
the actual header length in many cases.

As a result, many logs get shifted by 4 bytes. Commands were manually tested to
determine if they were affected or not. A test was added when reading the header
to determine the message id before reading the full header so that the correct
number of bytes is read.

A report was filed with Novatel, who may fix this in the future.