Update output format and handle time syncing.
All cameras for a given timestep are printed on a single line. This allows
the alignment of the camera frames to be known without any postprocessing. A
simple test of the camera timestamps is performed to check for misalignment.

The initial offset between the camera clocks is found and if the offset in
subsequent frames differs by more than MAXOFFSET, an error is reported (a 1 in
the last column).

No error correction is performed and the clock drift is assumed negligible.