One camera per grabframe process.
This updates fixes a problem where when multiple cameras are used in a grabframe
process, there are duplicates frames. For example, the output for 10 frames will
update the frame number, time, etc, but the actual image that is saved is the
same for all 10 frames.

The exact cause is not known, but running the cameras in separate grabframe
processes appears to eliminate the problem.