User specified directory and camname.
In order to better handle the file placement for stereo cameras running in
separate grabframe processes, the user must now specify the directory and
camera name instead of creating these automatically as before.

Example:
$ grabframe foo bar

will save images into ./images/foo/bar

mkdir() was called in such a way that will prevent the user from overwriting a
directory with the same camera name in the same directory.