Fixed the error of duplicate frames.
The ImageFile struct was not being set correctly. ppcImageMem and pnImageID
were left NULL causing the API to return the latest image frame rather than the
current frame in the queue. By assigning the values associated with the current
location in the queue to these members, the correct frames are saved.