Use selfgenerating numbering for filenames, bad image handling.
The camera frame numbers are useful but cannot be trusted to be unique since they reset under certain circumstances. We generate our own number and map it to the camera frame number in the logs.

When an image capture error is detected, we no longer continue with the getimage routine, we just read the error message and return without saving.